Amend the commit user due to CLA signature, test=develop
test=develop
